Extract InMemorySink into dedicated Storage.InMemory project
Move InMemorySink and InMemorySinkFactory out of the core Storage
assembly into a new Cratis.Chronicle.Storage.InMemory project
(namespace Cratis.Chronicle.Storage.InMemory.Sinks), the foundation for
formalizing InMemory as a first-class, configurable storage backend
alongside MongoDB and SQL.

- Add Storage.InMemory and Storage.InMemory.Specs projects, registered
  in the solution.
- Relocate the InMemorySink specs into Storage.InMemory.Specs and fix
  their namespaces; carry over the pre-existing exclusion of the
  find-root-key specs that target a removed changeset API.
- Repoint the Testing client and the MongoDB sink-parity spec to the
  new sink location.
