+++ Source/Kernel/Storage.InMemory.Specs/Sinks/for_InMemorySink/when_applying_changes/and_change_removes_a_child.cs
@@ -10,54 +10,41 @@ using Cratis.Chronicle.Concepts.Sinks;
 using Cratis.Chronicle.Properties;
 using Cratis.Chronicle.Schemas;
 
-namespace Cratis.Chronicle.Storage.Sinks.InMemory.for_InMemorySink.when_applying_changes;
+namespace Cratis.Chronicle.Storage.InMemory.Sinks.for_InMemorySink.when_applying_changes;
 
-public class and_property_change_targets_the_same_collection_as_child_added : Specification
+public class and_change_removes_a_child : Specification
 {
+    const string RemovedRoleId = "role-1";
+    const string KeptRoleId = "role-2";
+
     InMemorySink _sink;
     IChangeset<AppendedEvent, ExpandoObject> _changeset;
     Key _key;
     ExpandoObject? _result;
-    IEnumerable<FailedPartition> _failedPartitions;
 
     void Establish()
     {
         _key = new Key("user-1", ArrayIndexers.NoIndexers);
         _sink = new InMemorySink(CreateReadModelDefinition(), new TypeFormats());
 
-        var role = CreateRole("Administrator");
-        var rolesProperty = new PropertyPath("roles");
-        var childAdded = new ChildAdded(role, rolesProperty, new PropertyPath("role"), "Administrator", ArrayIndexers.NoIndexers);
+        dynamic initialState = new ExpandoObject();
+        initialState.roles = new[] { CreateRole(RemovedRoleId), CreateRole(KeptRoleId) };
 
-        dynamic state = new ExpandoObject();
-        state.name = "Ada Lovelace";
-        state.roles = new[] { role };
-
-        PropertyDifference[] differences =
-        [
-            new PropertyDifference(new PropertyPath("name"), null, "Ada Lovelace"),
-            new PropertyDifference(rolesProperty, Array.Empty<object>(), new[] { role })
-        ];
-        var propertiesChanged = new PropertiesChanged<ExpandoObject>(state, differences);
+        var childRemoved = new ChildRemoved(CreateRole(RemovedRoleId), new PropertyPath("roles"), new PropertyPath("id"), RemovedRoleId);
 
         _changeset = Substitute.For<IChangeset<AppendedEvent, ExpandoObject>>();
-        _changeset.InitialState.Returns(new ExpandoObject());
-        Change[] changes = [propertiesChanged, childAdded];
-        _changeset.Changes.Returns(changes);
+        _changeset.InitialState.Returns((ExpandoObject)initialState);
+        _changeset.Changes.Returns([childRemoved]);
     }
 
     async Task Because()
     {
-        _failedPartitions = await _sink.ApplyChanges(_key, _changeset, 42UL);
+        await _sink.ApplyChanges(_key, _changeset, 42UL);
         _result = await _sink.FindOrDefault(_key);
     }
 
-    [Fact] void should_not_fail() => _failedPartitions.ShouldBeEmpty();
-    [Fact] void should_keep_non_conflicting_property_changes() => GetValue<string>("name").ShouldEqual("Ada Lovelace");
-    [Fact] void should_apply_the_child_operation_once() => GetRoles().Length.ShouldEqual(1);
-    [Fact] void should_keep_the_added_role() => GetValue<string>(GetRoles()[0], "role").ShouldEqual("Administrator");
-
-    T GetValue<T>(string property) => GetValue<T>(_result, property);
+    [Fact] void should_have_one_role_left() => GetRoles().Length.ShouldEqual(1);
+    [Fact] void should_keep_the_other_role() => GetValue<string>(GetRoles()[0], "id").ShouldEqual(KeptRoleId);
 
     static T GetValue<T>(ExpandoObject target, string property)
     {
@@ -67,14 +54,14 @@ public class and_property_change_targets_the_same_collection_as_child_added : Sp
 
     ExpandoObject[] GetRoles()
     {
-        var dictionary = (IDictionary<string, object?>)_result;
+        var dictionary = (IDictionary<string, object?>)_result!;
         return ((IEnumerable<object>)dictionary["roles"]).Cast<ExpandoObject>().ToArray();
     }
 
-    static ExpandoObject CreateRole(string role)
+    static ExpandoObject CreateRole(string id)
     {
         dynamic item = new ExpandoObject();
-        item.role = role;
+        item.id = id;
         return item;
     }
