+++ Source/Kernel/Storage.InMemory.Specs/Sinks/for_InMemorySink/when_trying_to_find_root_key_by_child_value/and_child_value_does_not_exist.cs
@@ -4,7 +4,9 @@
 using Cratis.Chronicle.Concepts.Keys;
 using Cratis.Chronicle.Properties;
 
-namespace Cratis.Chronicle.Storage.Sinks.InMemory.for_InMemorySink.when_trying_to_find_root_key_by_child_value;
+using DomainChanges = Cratis.Chronicle.Changes;
+
+namespace Cratis.Chronicle.Storage.InMemory.Sinks.for_InMemorySink.when_trying_to_find_root_key_by_child_value;
 
 public class and_child_value_does_not_exist : given.an_in_memory_sink
 {
@@ -18,13 +20,13 @@ public class and_child_value_does_not_exist : given.an_in_memory_sink
         _nonExistentId = Guid.NewGuid();
 
         // Insert a document but search for a non-existent child
-        await _sink.ApplyChanges(new Key(_rootKey, ArrayIndexers.NoIndexers), new Changes.Changeset(
+        await _sink.ApplyChanges(new Key(_rootKey, ArrayIndexers.NoIndexers), new DomainChanges.Changeset(
             [
-                new Changes.PropertiesChanged<object>(new Dictionary<PropertyPath, object>
+                new DomainChanges.PropertiesChanged<object>(new Dictionary<PropertyPath, object>
                 {
                     [new PropertyPath("name")] = "Root Document"
                 }),
-                new Changes.ChildAdded(
+                new DomainChanges.ChildAdded(
                     new PropertyPath("children"),
                     new Key(Guid.NewGuid(), ArrayIndexers.NoIndexers),
                     new Dictionary<PropertyPath, object>
